Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    float a,b,c;
+    int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
     printf("Average: %.2f",(a+b+c)/3.0);
     return 0;
